Move cnu code into CircPool and MatrixAnalysis
