+++ tket/src/Circuit/include/Circuit/CircPool.hpp
@@ -392,6 +392,28 @@ Circuit tk1_to_rzsx(const Expr &alpha, const Expr &beta, const Expr &gamma);
 
 Circuit tk1_to_tk1(const Expr &alpha, const Expr &beta, const Expr &gamma);
 
+/**
+ * @brief Get an n-qubit incrementer circuit with linear depth and O(n^2) gate
+ * count. There exists a global phase difference
+ * https://arxiv.org/abs/2203.11882
+ *
+ * @param n number of qubits
+ * @param lsb set to false if we don't want to toggle the least siginificant bit
+ * @return Circuit containing CRx, X
+ */
+Circuit incrementer_linear_depth(unsigned n, bool lsb = true);
+
+/**
+ * @brief Implement CnU gate with linear depth and O(n^2) gate count.
+ * https://arxiv.org/abs/2203.11882
+ *
+ * @param n number of controls
+ * @param u the controlled 2x2 unitary matrix
+ * @return Circuit containing CRx, Unitary1QBox, and one-qubit controlled
+ * Unitary1QBox
+ */
+Circuit cnu_linear_depth_decomp(unsigned n, const Eigen::Matrix2cd &u);
+
 }  // namespace CircPool
 
 }  // namespace tket
